requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.5.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
